requirements.txt
filterpy
yfinance
streamlit
pandas
numpy
PyYAML
matplotlib
